# [1/2] sgl-kernel: Fuse routed scaling factor into select_experts