<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

Follow up to https://github.com/sgl-project/sglang/pull/8333
Fuse the multiply by routed_scaling_factor into select_experts, following example of TRT-LLM:
https://github.com/NVIDIA/TensorRT-LLM/blob/main/tensorrt_llm/_torch/models/modeling_deepseekv3.py#L323
https://github.com/NVIDIA/TensorRT-LLM/blob/738ab615930fd08dccb94fa388bd74dc91c5f235/cpp/tensorrt_llm/kernels/noAuxTcKernels.cu#L651

For the non-FP4 paths, the routed_scaling_factor is fused into moe_sum_reduce. However, we could move it into select_experts for those paths too if we wanted to simplify the code.

## Modifications

Add boolean argument to fused_moe_gate()


## Results

Prefill:
10.46% speedup at BS 1
1.86% speedup at BS 128
Decode:
1.22% speedup at BS1
0.26% speedup at BS128

Server command
```
python3 -m sglang.launch_server --model-path nvidia/DeepSeek-R1-0528-FP4 --trust-remote-code --quantization modelopt_fp4 --tp 8 --enable-flashinfer-cutlass-moe --enable-ep-moe --ep-size 8
```

BEFORE results
```
python3 -m sglang.bench_one_batch_server --model-path nvidia/DeepSeek-R1-0528-FP4 --base-url http://127.0.0.1:30000/ --batch-size 1 --input-len 1024 --output-len 1024

#Input tokens: 1024
#Output tokens: 1024
batch size: 1
input_len: 1024
output_len: 1024
latency: 10.60 s
ttft: 0.11 s
last generation throughput: 96.61 tok/s
input throughput: 9392.46 tok/s
output throughput: 97.62 tok/s

python3 -m sglang.bench_one_batch_server --model-path nvidia/DeepSeek-R1-0528-FP4 --base-url http://127.0.0.1:30000/ --batch-size 128 --input-len 1024 --output-len 1024

#Input tokens: 131072
#Output tokens: 131072
batch size: 128
input_len: 1024
output_len: 1024
latency: 39.79 s
ttft: 7.56 s
last generation throughput: 3693.98 tok/s
input throughput: 17344.15 tok/s
output throughput: 4066.24 tok/s
```

AFTER results
```
python3 -m sglang.bench_one_batch_server --model-path nvidia/DeepSeek-R1-0528-FP4 --base-url http://127.0.0.1:30000/ --batch-size 1 --input-len 1024 --output-len 1024

#Input tokens: 1024
#Output tokens: 1024
batch size: 1
input_len: 1024
output_len: 1024
latency: 10.46 s
ttft: 0.10 s
last generation throughput: 97.88 tok/s
input throughput: 10375.70 tok/s
output throughput: 98.82 tok/s

python3 -m sglang.bench_one_batch_server --model-path nvidia/DeepSeek-R1-0528-FP4 --base-url http://127.0.0.1:30000/ --batch-size 128 --input-len 1024 --output-len 1024

#Input tokens: 131072
#Output tokens: 131072
batch size: 128
input_len: 1024
output_len: 1024
latency: 39.57 s
ttft: 7.42 s
last generation throughput: 3698.44 tok/s
input throughput: 17667.47 tok/s
output throughput: 4076.75 tok/s
```


Accuracy
```
python3 benchmark/gsm8k/bench_sglang.py --num-shots 8 --num-questions 1319 --parallel 1319 --port=30000

# Before
Accuracy: 0.960
Invalid: 0.000
Latency: 23.902 s
Output throughput: 6057.485 token/s

# After
Accuracy: 0.961
Invalid: 0.000
Latency: 23.413 s
Output throughput: 6128.773 token/s
```

## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.