@trevor-m 2025-08-01
SGL_CUTLASS_MOE=1 FP8 path was giving 0.832 accuracy because we weren't multiplying by routed_scaling_Factor. I enabled the fusion to fix it and now it gives 0.965.